[ATR-227] refactor: s3 url prefix 제거 (#11)

+++ src/main/kotlin/attraction/run/s3/S3Service.kt
@@ -3,6 +3,7 @@ package attraction.run.s3
 import attraction.run.article.Article
 import io.awspring.cloud.s3.ObjectMetadata
 import io.awspring.cloud.s3.S3Operations
+import io.awspring.cloud.s3.S3Resource
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Service
 import org.springframework.util.MimeTypeUtils.TEXT_HTML_VALUE
@@ -24,22 +25,21 @@ class S3Service(
         }
     }
 
-    fun upload(articles: List<Article>): List<Article> {
-        return articles.onEach(::articleUpload)
+    fun uploadAllArticle(articles: List<Article>): List<Article> {
+        return articles.onEach(::uploadArticle)
     }
 
-    private fun articleUpload(article: Article): Article {
+    private fun uploadArticle(article: Article): Article {
         val file = createFileFromHTMLContent(article.contentHTML)
+        val upload = uploadFileToS3(file)
 
-        val upload = s3Operations.upload(bucket, file.name, file.inputStream(),
-                ObjectMetadata.builder()
-                        .contentType(TEXT_HTML_VALUE)
-                        .build())
-        val uploadUrl = upload.url.toString()
+        if (!uploadUrlContainsFileName(upload, file)) {
+            throw IllegalArgumentException("파일 이름이 올바르지 않습니다.")
+        }
 
         return article.apply {
             file.delete()
-            this.contentUrl = uploadUrl
+            this.contentUrl = file.name
         }
     }
 
@@ -47,7 +47,20 @@ class S3Service(
         val filePath = "$path/${UUID.randomUUID()}.html"
         val file = File(filePath)
 
-        file.writeText(htmlContent)
-        return file
+        return file.apply {
+            writeText(htmlContent)
+        }
     }
+
+    private fun uploadFileToS3(file: File) = s3Operations
+            .upload(
+                    bucket,
+                    file.name,
+                    file.inputStream(),
+                    ObjectMetadata.builder()
+                            .contentType(TEXT_HTML_VALUE)
+                            .build()
+            )
+
+    private fun uploadUrlContainsFileName(upload: S3Resource, file: File) = upload.url.toString().contains(file.name)
 }
